+++ src/kana/newKatakanaCombinations.js
@@ -1,0 +1,41 @@
+export const newKatakanaCombinations = [
+    { word: "ファ", answer: "fa" },
+    { word: "フィ", answer: "fi" },
+    { word: "フェ", answer: "fe" },
+    { word: "フォ", answer: "fo" },
+
+    { word: "ヴァ", answer: "va" },
+    { word: "ヴィ", answer: "vi" },
+    { word: "ヴ", answer: "vu" },
+    { word: "ヴェ", answer: "ve" },
+    { word: "ヴォ", answer: "vo" },
+
+    { word: "ウィ", answer: "wi" },
+    { word: "ウェ", answer: "we" },
+    { word: "ウォ", answer: "wo" },
+
+    { word: "シェ", answer: "she" },
+    { word: "ジェ", answer: "je" },
+    { word: "チェ", answer: "che" },
+
+    { word: "ティ", answer: "ti" },
+    { word: "ディ", answer: "di" },
+    { word: "トゥ", answer: "tu" },
+    { word: "ドゥ", answer: "du" },
+
+    { word: "ツァ", answer: "tsa" },
+    { word: "ツィ", answer: "tsi" },
+    { word: "ツェ", answer: "tse" },
+    { word: "ツォ", answer: "tso" },
+
+    { word: "スィ", answer: "si" },
+    { word: "ズィ", answer: "zi" },
+    { word: "デュ", answer: "dyu" },
+    { word: "テュ", answer: "tyu" },
+    { word: "クォ", answer: "kwo" },
+];
+
+export default {
+    list: newKatakanaCombinations,
+    ref: "https://www.youtube.com/watch?v=1vmzhmTobXc&t=0"
+}