+++ src/vocabulary/index.js
@@ -6,10 +6,10 @@ import numbers from "./numbers";
 import politeness from "./politeness";
 
 export default {
-    "Les Formules de Politesse": politeness,
     "Premier Mots en Kanji": firstKanji,
-    "Les Nombres": numbers,
+    "Les Formules de Politesse": politeness,
     "Le Matériel Domestique": domesticEquipment,
     "Le Nom des Pays": countryNames,
     "Les Animaux Domestiques": domesticAnimals,
+    "Les Nombres": numbers,
 };